move option content to subcomponent
