+++ packages/combobox/package.json
@@ -26,6 +26,7 @@
     "@leafygreen-ui/icon": "^11.12.4",
     "@leafygreen-ui/icon-button": "^15.0.4",
     "@leafygreen-ui/inline-definition": "^6.0.2",
+    "@leafygreen-ui/input-option": "^0.1.0",
     "@leafygreen-ui/lib": "^10.1.0",
     "@leafygreen-ui/palette": "^3.4.7",
     "@leafygreen-ui/popover": "^11.0.5",
