+++ packages/combobox/src/ComboboxOption/ComboboxOption.tsx
@@ -1,30 +1,21 @@
-import React, { useCallback, useContext, useMemo } from 'react';
+import React, { useCallback, useContext } from 'react';
 import PropTypes from 'prop-types';
 
-import Checkbox from '@leafygreen-ui/checkbox';
-import { css, cx } from '@leafygreen-ui/emotion';
-import { useForwardedRef, useIdAllocator } from '@leafygreen-ui/hooks';
-import Icon, { isComponentGlyph } from '@leafygreen-ui/icon';
-import { palette } from '@leafygreen-ui/palette';
+import { cx } from '@leafygreen-ui/emotion';
+import { useForwardedRef } from '@leafygreen-ui/hooks';
+import { InputOption } from '@leafygreen-ui/input-option';
 
 import {
   ComboboxOptionProps,
   InternalComboboxOptionProps,
 } from '../Combobox.types';
 import { ComboboxContext } from '../ComboboxContext';
-import { wrapJSX } from '../utils';
 
 import {
-  checkIconStyle,
-  comboboxOptionActiveStyle,
   comboboxOptionBaseStyle,
-  comboboxOptionDisabledStyle,
   comboboxOptionSizeStyle,
-  comboboxOptionThemeStyle,
-  disallowPointer,
-  displayNameStyle,
-  flexSpan,
 } from './ComboboxOption.styles';
+import { OptionContent } from './OptionContent';
 
 /**
  * Component
@@ -35,19 +26,18 @@ export const InternalComboboxOption = React.forwardRef<
 >(
   (
     {
-      displayName,
       glyph,
       isSelected,
+      displayName,
       isFocused,
       disabled,
       setSelected,
       className,
+      ...rest
     }: InternalComboboxOptionProps,
     forwardedRef,
   ) => {
-    const { multiselect, darkMode, theme, withIcons, inputValue, size } =
-      useContext(ComboboxContext);
-    const optionTextId = useIdAllocator({ prefix: 'combobox-option-text' });
+    const { darkMode, size } = useContext(ComboboxContext);
     const optionRef = useForwardedRef(forwardedRef, null);
 
     const handleOptionClick = useCallback(
@@ -70,104 +60,30 @@ export const InternalComboboxOption = React.forwardRef<
       [disabled, optionRef, setSelected],
     );
 
-    const renderedIcon = useMemo(() => {
-      if (glyph) {
-        if (isComponentGlyph(glyph)) {
-          return glyph;
-        }
-        console.error(
-          '`ComboboxOption` instance did not render icon because it is not a known glyph element.',
-          glyph,
-        );
-      }
-    }, [glyph]);
-
-    const renderedChildren = useMemo(() => {
-      // Multiselect
-      if (multiselect) {
-        const checkbox = (
-          <Checkbox
-            aria-labelledby={optionTextId}
-            checked={isSelected}
-            tabIndex={-1}
-            disabled={disabled}
-            darkMode={darkMode}
-            className={css`
-              // TODO: Remove when this is resolved:
-              // https://jira.mongodb.org/browse/PD-2171
-              & > label > div {
-                margin-top: 0;
-              }
-            `}
-          />
-        );
-
-        return (
-          <>
-            <span className={cx(flexSpan, disallowPointer)}>
-              {withIcons ? renderedIcon : checkbox}
-              <span id={optionTextId} className={displayNameStyle(isSelected)}>
-                {wrapJSX(displayName, inputValue, 'strong')}
-              </span>
-            </span>
-            {withIcons && checkbox}
-          </>
-        );
-      }
-
-      // Single select
-      return (
-        <>
-          <span className={cx(flexSpan, disallowPointer)}>
-            {renderedIcon}
-            <span className={displayNameStyle(isSelected)}>
-              {wrapJSX(displayName, inputValue, 'strong')}
-            </span>
-          </span>
-          {isSelected && (
-            <Icon
-              glyph="Checkmark"
-              className={checkIconStyle[size]}
-              color={darkMode ? palette.blue.light1 : palette.blue.base}
-            />
-          )}
-        </>
-      );
-    }, [
-      multiselect,
-      renderedIcon,
-      isSelected,
-      displayName,
-      inputValue,
-      size,
-      darkMode,
-      optionTextId,
-      disabled,
-      withIcons,
-    ]);
-
     return (
-      <li
+      <InputOption
+        {...rest}
+        as="li"
         ref={optionRef}
-        role="option"
-        aria-selected={isFocused}
+        focused={isFocused}
+        disabled={disabled}
         aria-label={displayName}
-        tabIndex={-1}
+        darkMode={darkMode}
         className={cx(
           comboboxOptionBaseStyle,
           comboboxOptionSizeStyle[size],
-          comboboxOptionThemeStyle[theme],
-          {
-            [comboboxOptionActiveStyle[theme]]: isFocused,
-            [comboboxOptionDisabledStyle[theme]]: disabled,
-          },
           className,
         )}
         onClick={handleOptionClick}
         onKeyDown={handleOptionClick}
       >
-        {renderedChildren}
-      </li>
+        <OptionContent
+          disabled={disabled}
+          displayName={displayName}
+          glyph={glyph}
+          isSelected={isSelected}
+        />
+      </InputOption>
     );
   },
 );
