+++ packages/combobox/tsconfig.json
@@ -28,6 +28,9 @@
       "path": "../inline-definition"
     },
     {
+      "path": "../input-option"
+    },
+    {
       "path": "../lib"
     },
     {
